MTM now uses separate IO file for gripper analog input.  This simplifies PID code, avoids strange numbers displayed in Qt IO widget, allows to remove NumberOfAxes, moved code re. analog inputs to MTM class.

+++ components/code/Qt/mtsIntuitiveResearchKitConsoleQt.cpp
@@ -127,7 +127,7 @@ void mtsIntuitiveResearchKitConsoleQt::Configure(mtsIntuitiveResearchKitConsole 
         case mtsIntuitiveResearchKitConsole::Arm::ARM_ECM:
         case mtsIntuitiveResearchKitConsole::Arm::ARM_ECM_DERIVED:
             // PID widget
-            unsigned int numberOfJoints;
+            size_t numberOfJoints;
             if ((armIter->second->mType == mtsIntuitiveResearchKitConsole::Arm::ARM_PSM) ||
                 (armIter->second->mType == mtsIntuitiveResearchKitConsole::Arm::ARM_PSM_DERIVED)) {
                 numberOfJoints = 7;
